requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
